Found bug
Hey Ryan, the problem turned out to be the result of an "else if" that should have only been an "if." The reason this broke the code is because the function can only enter/execute one branch of an if structure. So if either gridArray[i][i] == "X" or gridArray[i][(2-i)] == "O" is true, then there is NO chance that the function will even check if i == 2 (let alone enter that if branch.)

I commented out the trouble making elses and drew a shitty arrow that points to their location.